Refactor: rename components and related code
- NavLink -> NavMenu
- use original component names when importing
