+++ client/public/locales/de/navMenu.json
