+++ client/src/components/About.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { List } from 'semantic-ui-react';
 
 import Footer from './Footer';
-import NavMenu from './NavLink';
+import NavMenu from './NavMenu';
 
 const About = () => (
   <>
