+++ client/src/components/HeaderWithLink/Ask.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { withTranslation } from 'react-i18next';
 
-import AskADoctor from '../../containers/AddQuestionForm';
+import AddQuestionForm from '../../containers/AddQuestionForm';
 import '../../styles/Ask.css';
 
 const Ask = (props) => {
@@ -11,7 +11,7 @@ const Ask = (props) => {
         <p>{props.t('patientBoard:banner.cannotAnswer')}</p>
       </div>
       <div className="ask-button-wrapper">
-        <AskADoctor buttonLabel={props.buttonLabel} />
+        <AddQuestionForm buttonLabel={props.buttonLabel} />
       </div>
     </div>
   );
