+++ client/src/i18n.js
@@ -13,7 +13,7 @@ i18n
       loadPath: '/locales/{{lng}}/{{ns}}.json',
     },
     // string or array of namespaces to load
-    ns: ['common', 'navLink', 'patientBoard', 'physicianView'],
+    ns: ['common', 'navMenu', 'patientBoard', 'physicianView'],
     defaultNS: 'patientBoard',
     // array of allowed languages
     // @toDo add other languages
