+++ client/src/styles/NavMenu.css
@@ -57,12 +57,12 @@ nav ul li a:hover {
   opacity: 0.75;
 }
 
-.close-menu{
+.close-menu {
   text-align: right;
   font-size: 2em;
-  padding:10px 10px 0px 0px;
+  padding: 10px 10px 0px 0px;
   color: white;
-  opacity: 1.0;
+  opacity: 1;
 }
 .row#sponsors {
   margin-top: 500px;
@@ -112,5 +112,3 @@ button.ui.button.logInlogOut {
 .sidebar_language_dropdown {
   background-color: #1678c2;
 }
-
-
